backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
djoser
django-cors-headers
requests
